use azure-json

+++ sdk/openai/azure-ai-openai/src/samples/java/com/azure/ai/openai/impl/ReadmeSamples.java
@@ -217,7 +217,7 @@ public final class ReadmeSamples {
 
         for (EmbeddingItem item : embeddings.getData()) {
             System.out.printf("Index: %d.%n", item.getPromptIndex());
-            for (Float embedding : item.getEmbedding()) {
+            for (Double embedding : item.getEmbedding()) {
                 System.out.printf("%f;", embedding);
             }
         }
