+++ sdk/openai/azure-ai-openai/src/samples/java/com/azure/ai/openai/usage/GetEmbeddingsAsyncSample.java
@@ -40,7 +40,7 @@ public class GetEmbeddingsAsyncSample {
                     System.out.printf("Index: %d.%n", item.getPromptIndex());
                     System.out.println("Embedding as base64 encoded string: " +  item.getEmbeddingAsString());
                     System.out.println("Embedding as list of floats: ");
-                    for (Float embedding : item.getEmbedding()) {
+                    for (Double embedding : item.getEmbedding()) {
                         System.out.printf("%f;", embedding);
                     }
                 }
